feat: Finish impletation of product module

+++ src/modules/products/services/products.service.ts
@@ -1,31 +1,55 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { ProductsRepository } from '../repository/products.repository';
 import { ProductModel } from '../entities/product.model';
 import { Prisma } from '@prisma/client';
+import { CreateProductDto } from '../dto/create-product.dto';
+import { FindProductsDto } from '../dto/find-product.dto';
 
 @Injectable()
 export class ProductsService {
   constructor(private productsRepository: ProductsRepository) {}
 
-  create(data: ProductModel) {
-    return this.productsRepository.create(data);
+  async create(data: CreateProductDto) {
+    const product = new ProductModel(data);
+    await this.productsRepository.create(product);
+
+    return product.getEntity();
   }
 
-  findAll(params: Prisma.ProductWhereInput) {
+  findAll(params: FindProductsDto) {
     return this.productsRepository.findMany(params);
   }
 
-  findOne(id: number) {
-    return this.productsRepository.findOneOrNull({ id });
+  async findOne(id: number) {
+    const product = await this.productsRepository.findOneOrNull({ id });
+
+    if (!product) {
+      throw new BadRequestException('Product not found');
+    }
+
+    return product.getEntity();
   }
 
-  update(product: ProductModel) {
-    const id = product.getId();
-    const data = product.getEntity();
-    return this.productsRepository.update({ id }, data);
+  async update(id: number, data: Prisma.ProductUpdateInput) {
+    const productFound = await this.productsRepository.findOneOrNull({ id });
+    if (!productFound) {
+      throw new BadRequestException('Product not found');
+    }
+
+    const newProduct = await this.productsRepository.update(
+      { id: productFound.getId() },
+      data,
+    );
+
+    return newProduct.getEntity();
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const productFound = await this.productsRepository.findOneOrNull({ id });
+    if (!productFound) {
+      throw new BadRequestException('Product not found');
+    }
+
     return this.productsRepository.delete({ id });
   }
 }
